Maintain library search path order when beginning best version match checks.
A higher version in a later path will still win out. This change is designed to address the case where java.library.path has 64-bit and 32-bit lib paths, with the expectation that the 64-bit ones will be matched first. Without maintaining insertion order, it's possible for a 32-bit library to be found and loaded when a 64-bit one would have been earlier in the library load path.