Use Java 9 Thread constructor instead of private field. Fixes #105
The logic here now attempts to ways to stop threads from
inheriting thread locals:

* On Java 9+ it will use the new Thread constructor that takes
  a boolean for the sharing behavior.
* On Java 8- (when accessible) it will tweak an internal field on
  the Thread.

The fallback remains the same; if neither of the above will work,
we warn once. If either of the above fail at runtime, we warn
each time a Finalizer is created.

Note that this changes Finalizer from a Thread to a Runnable so
that Finalizer.start can lazily determine how to set up a non-
inheriting Thread.

All tests pass with this change on Java 8 and Java 9.