Start nested thread in the Finalizer. Fixes #116
The changes in 601bec2 converted the Finalizer from a Thread to a
Runnable which holds an internal reference to a new Thread.  This change
caused the run() method on the Finalizer to no longer be invoked,
leading to the cleanUp() it had been doing previously not being done
anymore.  In this commit, start() is called on the internal Thread.
This enables the Finalizer.run() method to be called on the new Thread,
where the cleanUp() would now start happening again.