Force compiler plugin version and add release flag
I'm not sure which version added support for the Java 9+ release
flag, but might as well go to latest. This patch will add the
"release 8" flag to ensure things like ByteBuffer.flip calls
compile to the Java 8-compatible version.

This eliminates the need for shims and manual casting to Buffer.