Pass array pinning flag down to jffi
Pinning primitive arrays enables access to their contents in native code
without requiring copying.

The contents of the array is accessed via the GetPrimitiveArrayCritical
JNI method. Different JVMs and different JVM configurations have
different implementations and consequences of this method. In G1 (on
Hotspot in OpenJDK 11) it currently acquires a lock that prevents
garbage collection ("GCLocker") and thus can negatively impact forward
progress of mutators if a collection is required. Native methods are
thus expected to not hold these arrays for long, as is documented in JNI
docs [1] and in the `@Pinned` javadocs.

In the case that a VM is unable for some reason to avoid a copy,
GetPrimitiveArrayCritical will return a copy. This will cause test
failure: I've included a test to verify that pinning 'works': this is
the best strategy I've come up with to verify that
GetPrimitiveArrayCritical is being called as a result of the param
annotation + flag. Access to some sort of counter, if available, might
be preferrable but I'm unaware any such portable and accessible
counters.

Fortunately, OpenJDK 11 and OpenJ9's JDK 11-compatible distribution both
seem to support pinned access, as does Dalvik/ART. Given that all the
popular implementations seem to support pinning, I've left the test
enabled.

Performance benefits of pinning is significant: the attached test
displays this, as does the example at
bkgood/jnr-ffi-array-pinning-tests. I embarked on this due to
surprisingly limited performance I got in a WIP libsnappy binding: with
copying (including judicious use of `@In` and `@Out`) I get about 90
MB/s but enabling array pinning gets it closer to 250 MB/s.

This was largely implemented in jnr/jffi@a61b1fc42aa7; the requisite
flag just needs to make it to the native stubs.

Fixes jnr/jnr-ffi#68.

[1]
https://docs.oracle.com/en/java/javase/11/docs/specs/jni/functions.html#getprimitivearraycritical-releaseprimitivearraycritical