Fix TypeAliases for aarch64 linux
`blksize_t` and `nlink_t` are incorrectly defined for aarch64 linux.

This currently manifests itself when using `File.stat` in jruby, where
the 'uid`, `gid`, and `nlink` values are incorrect.

The value of `nlink_t` is defined as follows:

```
/usr/include/bits/stat.h:    __nlink_t st_nlink;		/* Link count.  */
/usr/include/bits/types.h:__STD_TYPE __NLINK_T_TYPE __nlink_t;	/* Type of file link counts.  */
/usr/include/bits/typesizes.h:#define __NLINK_T_TYPE		__U32_TYPE
```

The value of `blksize_t` is defined as follows:

```
/usr/include/bits/stat.h:    __blksize_t st_blksize;	/* Optimal block size for I/O.  */
/usr/include/bits/types.h:__STD_TYPE __BLKSIZE_T_TYPE __blksize_t;
/usr/include/bits/typesizes.h:#define __BLKSIZE_T_TYPE	__S32_TYPE
```

This commit updates the type aliases for these types to match correctly.