Fix issue where string references aren't always completed with a null terminator.
This is particularly easy to reproduce with an empty string - the
existing implementation works most of the time because most characters
don't consume the full four bytes a single UTF-8 character could
consume.

The tests relies on the fact that two consecutive memory allocations
usually end up next to each other. This happens reliably for me on my
machine, but isn't guaranteed. If there's a better way to test this,
I'd love to hear it.